DRTII-407 Refactor persistAndMaybeSnapshot. Break out some classes
