+++ server/src/main/scala/actors/FixedPointsActorBase.scala
@@ -90,7 +90,7 @@ abstract class FixedPointsActorBase(now: () => SDateLike) extends RecoveryActorL
 
         val createdAt = now()
         val fixedPointsMessage = FixedPointsMessage(fixedPointsToFixedPointsMessages(state, createdAt), Option(createdAt.millisSinceEpoch))
-        persistAndMaybeSnapshotWithAck(fixedPointsMessage, sender(), SetFixedPointsAck(fixedPointStaffAssignments))
+        persistAndMaybeSnapshot(fixedPointsMessage, Option(sender(), SetFixedPointsAck(fixedPointStaffAssignments)))
       } else {
         log.info(s"No change. Nothing to persist")
         sender() ! SetFixedPointsAck(fixedPointStaffAssignments)
