+++ server/src/main/scala/actors/StaffMovementsActorBase.scala
@@ -146,7 +146,7 @@ class StaffMovementsActorBase(val now: () => SDateLike,
       log.info(s"Added $movementsToAdd. We have ${state.staffMovements.movements.length} movements after purging")
       val movements: StaffMovements = StaffMovements(movementsToAdd)
       val messagesToPersist = StaffMovementsMessage(staffMovementsToStaffMovementMessages(movements), Option(now().millisSinceEpoch))
-      persistAndMaybeSnapshotWithAck(messagesToPersist, sender(), AddStaffMovementsAck(movementsToAdd))
+      persistAndMaybeSnapshot(messagesToPersist, Option(sender(), AddStaffMovementsAck(movementsToAdd)))
 
     case RemoveStaffMovements(uuidToRemove) =>
       val updatedStaffMovements = state.staffMovements - Seq(uuidToRemove)
@@ -154,7 +154,7 @@ class StaffMovementsActorBase(val now: () => SDateLike,
 
       log.info(s"Removed $uuidToRemove. We have ${state.staffMovements.movements.length} movements after purging")
       val messagesToPersist: RemoveStaffMovementMessage = RemoveStaffMovementMessage(Option(uuidToRemove.toString), Option(now().millisSinceEpoch))
-      persistAndMaybeSnapshotWithAck(messagesToPersist, sender(), RemoveStaffMovementsAck(uuidToRemove))
+      persistAndMaybeSnapshot(messagesToPersist, Option(sender(), RemoveStaffMovementsAck(uuidToRemove)))
 
     case SaveSnapshotSuccess(md) =>
       log.info(s"Save snapshot success: $md")
